Fix for iOS mobile embed link on the REPL cog
Fixes an iOS mobile bug where links wrapped in ticks don't work - this removes them, making it clickable. Regular embed links work fine (that bug was fixed), but not bold links or any links wrapped in ticks that are inside embeds. This fixes it so the Hastebin URL is clickable on iOS devices using the mobile app with the repl/embed shell.